+++ packages/block-editor/src/hooks/align.js
@@ -6,7 +6,6 @@ import classnames from 'classnames';
 /**
  * WordPress dependencies
  */
-import { createHigherOrderComponent } from '@wordpress/compose';
 import { addFilter } from '@wordpress/hooks';
 import {
 	getBlockSupport,
@@ -98,8 +97,8 @@ export function addAttribute( settings ) {
 			...settings.attributes,
 			align: {
 				type: 'string',
-				// Allow for '' since it is used by updateAlignment function
-				// in withToolbarControls for special cases with defined default values.
+				// Allow for '' since it is used by the `updateAlignment` function
+				// in toolbar controls for special cases with defined default values.
 				enum: [ ...ALL_ALIGNMENTS, '' ],
 			},
 		};
@@ -108,14 +107,14 @@ export function addAttribute( settings ) {
 	return settings;
 }
 
-function BlockEditAlignmentToolbarControls( {
-	blockName,
-	attributes,
+function BlockEditAlignmentToolbarControlsPure( {
+	name: blockName,
+	align,
 	setAttributes,
 } ) {
 	// Compute the block valid alignments by taking into account,
 	// if the theme supports wide alignments or not and the layout's
-	// availble alignments. We do that for conditionally rendering
+	// available alignments. We do that for conditionally rendering
 	// Slot.
 	const blockAllowedAlignments = getValidAlignments(
 		getBlockSupport( blockName, 'align' ),
@@ -144,7 +143,7 @@ function BlockEditAlignmentToolbarControls( {
 	return (
 		<BlockControls group="block" __experimentalShareWithChildBlocks>
 			<BlockAlignmentControl
-				value={ attributes.align }
+				value={ align }
 				onChange={ updateAlignment }
 				controls={ validAlignments }
 			/>
@@ -152,79 +151,30 @@ function BlockEditAlignmentToolbarControls( {
 	);
 }
 
-/**
- * Override the default edit UI to include new toolbar controls for block
- * alignment, if block defines support.
- *
- * @param {Function} BlockEdit Original component.
- *
- * @return {Function} Wrapped component.
- */
-export const withToolbarControls = createHigherOrderComponent(
-	( BlockEdit ) => ( props ) => {
-		const hasAlignmentSupport = hasBlockSupport(
-			props.name,
-			'align',
-			false
-		);
-
-		return (
-			<>
-				{ hasAlignmentSupport && (
-					<BlockEditAlignmentToolbarControls
-						blockName={ props.name }
-						attributes={ props.attributes }
-						setAttributes={ props.setAttributes }
-					/>
-				) }
-				<BlockEdit key="edit" { ...props } />
-			</>
-		);
+export default {
+	shareWithChildBlocks: true,
+	edit: BlockEditAlignmentToolbarControlsPure,
+	useBlockProps,
+	addSaveProps: addAssignedAlign,
+	attributeKeys: [ 'align' ],
+	hasSupport( name ) {
+		return hasBlockSupport( name, 'align', false );
 	},
-	'withToolbarControls'
-);
+};
 
-function BlockListBlockWithDataAlign( { block: BlockListBlock, props } ) {
-	const { name, attributes } = props;
-	const { align } = attributes;
+function useBlockProps( { name, align } ) {
 	const blockAllowedAlignments = getValidAlignments(
 		getBlockSupport( name, 'align' ),
 		hasBlockSupport( name, 'alignWide', true )
 	);
 	const validAlignments = useAvailableAlignments( blockAllowedAlignments );
 
-	let wrapperProps = props.wrapperProps;
 	if ( validAlignments.some( ( alignment ) => alignment.name === align ) ) {
-		wrapperProps = { ...wrapperProps, 'data-align': align };
+		return { 'data-align': align };
 	}
 
-	return <BlockListBlock { ...props } wrapperProps={ wrapperProps } />;
+	return {};
 }
-
-/**
- * Override the default block element to add alignment wrapper props.
- *
- * @param {Function} BlockListBlock Original component.
- *
- * @return {Function} Wrapped component.
- */
-export const withDataAlign = createHigherOrderComponent(
-	( BlockListBlock ) => ( props ) => {
-		// If an alignment is not assigned, there's no need to go through the
-		// effort to validate or assign its value.
-		if ( props.attributes.align === undefined ) {
-			return <BlockListBlock { ...props } />;
-		}
-
-		return (
-			<BlockListBlockWithDataAlign
-				block={ BlockListBlock }
-				props={ props }
-			/>
-		);
-	},
-	'withDataAlign'
-);
 
 /**
  * Override props assigned to save component to inject alignment class name if
@@ -257,21 +207,6 @@ export function addAssignedAlign( props, blockType, attributes ) {
 
 addFilter(
 	'blocks.registerBlockType',
-	'core/align/addAttribute',
+	'core/editor/align/addAttribute',
 	addAttribute
-);
-addFilter(
-	'editor.BlockListBlock',
-	'core/editor/align/with-data-align',
-	withDataAlign
-);
-addFilter(
-	'editor.BlockEdit',
-	'core/editor/align/with-toolbar-controls',
-	withToolbarControls
-);
-addFilter(
-	'blocks.getSaveContent.extraProps',
-	'core/align/addAssignedAlign',
-	addAssignedAlign
 );