+++ packages/edit-site/src/components/sidebar-dataviews/index.js
@@ -2,65 +2,56 @@
  * WordPress dependencies
  */
 import { __experimentalItemGroup as ItemGroup } from '@wordpress/components';
-import { page, columns } from '@wordpress/icons';
-import { privateApis as routerPrivateApis } from '@wordpress/router';
 
+import { privateApis as routerPrivateApis } from '@wordpress/router';
 /**
  * Internal dependencies
  */
-import { useLink } from '../routes/link';
-import { default as DEFAULT_VIEWS } from '../page-pages/default-views';
+
+import { default as DEFAULT_VIEWS } from './default-views';
 import { unlock } from '../../lock-unlock';
 const { useLocation } = unlock( routerPrivateApis );
-import SidebarNavigationItem from '../sidebar-navigation-item';
+import DataViewItem from './dataview-item';
+import CustomDataViewsList from './custom-dataviews-list';
 
-function getDataViewIcon( dataview ) {
-	const icons = { list: page, grid: columns };
-	return icons[ dataview.view.type ];
-}
-
-function DataViewItem( { dataview, isActive, icon } ) {
-	const {
-		params: { path },
-	} = useLocation();
-
-	const _icon = icon || getDataViewIcon( dataview );
-
-	const linkInfo = useLink( {
-		path,
-		activeView: dataview.slug,
-	} );
-	return (
-		<SidebarNavigationItem
-			icon={ _icon }
-			{ ...linkInfo }
-			aria-current={ isActive ? 'true' : undefined }
-		>
-			{ dataview.title }
-		</SidebarNavigationItem>
-	);
-}
+const PATH_TO_TYPE = {
+	'/pages': 'page',
+};
 
 export default function DataViewsSidebarContent() {
 	const {
-		params: { path, activeView = 'all' },
+		params: { path, activeView = 'all', isCustom = 'false' },
 	} = useLocation();
-	if ( ! path || path !== '/pages' ) {
+	if ( ! path || ! PATH_TO_TYPE[ path ] ) {
 		return null;
 	}
+	const type = PATH_TO_TYPE[ path ];
 
 	return (
-		<ItemGroup>
-			{ DEFAULT_VIEWS.map( ( dataview ) => {
-				return (
-					<DataViewItem
-						key={ dataview.slug }
-						icon={ dataview.icon }
-						dataview={ dataview }
-						isActive={ dataview.slug === activeView }
-					/>
-				);
-			} ) }
-		</ItemGroup>
+		<>
+			<ItemGroup>
+				{ DEFAULT_VIEWS[ type ].map( ( dataview ) => {
+					return (
+						<DataViewItem
+							key={ dataview.slug }
+							slug={ dataview.slug }
+							title={ dataview.title }
+							icon={ dataview.icon }
+							type={ dataview.view.type }
+							isActive={
+								isCustom === 'false' &&
+								dataview.slug === activeView
+							}
+							isCustom="false"
+						/>
+					);
+				} ) }
+			</ItemGroup>
+			<CustomDataViewsList
+				activeView={ activeView }
+				type={ type }
+				isCustom="true"
+			/>
+		</>
 	);
 }