+++ packages/react-native-editor/__device-tests__/gutenberg-editor-drag-and-drop.test.js
@@ -2,14 +2,10 @@
  * Internal dependencies
  */
 import { blockNames } from './pages/editor-page';
-import { clearClipboard, dragAndDropAfterElement } from './helpers/utils';
+import { dragAndDropAfterElement } from './helpers/utils';
 import testData from './helpers/test-data';
 
 describe( 'Gutenberg Editor Drag & Drop blocks tests', () => {
-	beforeEach( async () => {
-		await clearClipboard( editorPage.driver );
-	} );
-
 	it( 'should be able to drag & drop a block', async () => {
 		// Initialize the editor with a Spacer and Paragraph block
 		await editorPage.initializeEditor( {