+++ packages/scripts/scripts/test-playwright.js
@@ -12,6 +12,7 @@ process.on( 'unhandledRejection', ( err ) => {
 /**
  * External dependencies
  */
+const path = require( 'path' );
 const { resolve } = require( 'node:path' );
 const { sync: spawn } = require( 'cross-spawn' );
 
@@ -23,18 +24,28 @@ const {
 	hasProjectFile,
 	hasArgInCLI,
 	getArgsFromCLI,
+	getAsBooleanFromENV,
 } = require( '../utils' );
 
-const result = spawn(
-	'node',
-	[ require.resolve( 'playwright-core/cli' ), 'install' ],
-	{
-		stdio: 'inherit',
-	}
-);
+if ( ! getAsBooleanFromENV( 'PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD' ) ) {
+	const result = spawn(
+		'node',
+		[
+			path.resolve(
+				require.resolve( 'playwright-core' ),
+				'..',
+				'cli.js'
+			),
+			'install',
+		],
+		{
+			stdio: 'inherit',
+		}
+	);
 
-if ( result.status > 0 ) {
-	process.exit( result.status );
+	if ( result.status > 0 ) {
+		process.exit( result.status );
+	}
 }
 
 const config =